+++ inc/console.h
@@ -7,7 +7,7 @@
 #include <stdint.h>
 
 /* Put character at (x,y) with attribute */
-void console_putch_xy(uint32_t x, uint32_t y, uint8_t ch, uint8_t attr);
+void console_putch_xy(uint32_t x, uint32_t y, uint16_t ch, uint8_t attr);
 
 /* Write NUL-terminated string at (x,y) with attribute */
 void console_write_str_xy(uint32_t x, uint32_t y, const char *s, uint8_t attr);
@@ -19,7 +19,7 @@ void console_get_cursor(uint32_t *x, uint32_t *y);
 void console_begin_tty_batch(void);
 void console_end_tty_batch(void);
 /* Fill rectangle with character and attribute */
-void console_fill_rect(uint32_t x, uint32_t y, uint32_t w, uint32_t h, uint8_t ch, uint8_t attr);
+void console_fill_rect(uint32_t x, uint32_t y, uint32_t w, uint32_t h, uint16_t ch, uint8_t attr);
 
 /* Current console text geometry in character cells. */
 int console_max_rows(void);
@@ -31,5 +31,5 @@ void console_scroll_region_up(uint32_t top, uint32_t bottom, uint8_t attr);
 uint8_t console_get_cell_attr(uint32_t x, uint32_t y);
 
 /* Write one TTY cell from devfs: framebuffer path skips kernel ANSI state machines. */
-void console_putc_tty_literal(uint8_t ch, uint8_t attr);
+void console_putc_tty_literal(uint16_t ch, uint8_t attr);
